+++ src/CloudinaryEngine.php
@@ -9,14 +9,22 @@
 
 namespace CloudinaryLabs\CloudinaryLaravel;
 
+use Cloudinary\Api\Admin\AdminApi;
+use Cloudinary\Api\Search\SearchApi;
+use Cloudinary\Api\Upload\UploadApi;
+use Cloudinary\Asset\File;
+use Cloudinary\Asset\Image;
+use Cloudinary\Asset\Video;
 use Cloudinary\Cloudinary;
+use Cloudinary\Asset\Analytics;
+use Cloudinary\Configuration\Configuration;
 use Cloudinary\Tag\ImageTag;
 use Cloudinary\Tag\VideoTag;
 use Cloudinary\Api\ApiResponse;
 use Cloudinary\Api\BaseApiClient;
 use Cloudinary\Api\Exception\ApiError;
+use Exception;
 use GuzzleHttp\Promise\PromiseInterface;
-use CloudinaryLabs\Exceptions\IsNullException;
 
 /**
  * Class CloudinaryEngine
@@ -25,11 +33,11 @@ use CloudinaryLabs\Exceptions\IsNullException;
 class CloudinaryEngine
 {
     /**
-     * Laravel Package Version.
+     * Cloudinary Laravel Package Version.
      *
      * @var string PACKAGE_VERSION
      */
-    const PACKAGE_VERSION = '1.0.5';
+    const PACKAGE_VERSION = '2.0.2';
 
     public const ASSET_ID = 'asset_id';
     public const PUBLIC_ID = 'public_id';
@@ -56,23 +64,24 @@ class CloudinaryEngine
      * Instance of Cloudinary
      * @var Cloudinary
      */
-    protected $cloudinary;
+    protected Cloudinary $cloudinary;
 
     /**
-     * Instance of Cloudinary Config
-     * @var Configuration
+     * Cloudinary url
+     * @var string
      */
-    protected $cloudinaryConfig;
+    protected string $url;
 
     /**
      *  Response from Cloudinary
-     * @var Array
+     * @var array|ApiResponse
      */
-    protected $response;
+    protected array|ApiResponse $response;
 
     public function __construct()
     {
         $this->setUserPlatform();
+        $this->setAnalytics();
         $this->setCloudinaryConfig();
         $this->bootCloudinary();
     }
@@ -81,35 +90,44 @@ class CloudinaryEngine
      * Create a Cloudinary Config Instance
      *
      */
-    public function setCloudinaryConfig()
+    public function setCloudinaryConfig(): void
     {
-        $config = config('cloudinary.cloud_url');
-        $this->cloudinaryConfig = $config;
+        $this->url = config('cloudinary.cloud_url');
     }
 
     /**
      * Set User Agent and Platform
      *
      */
-    public function setUserPlatform()
+    public function setUserPlatform(): void
     {
         BaseApiClient::$userPlatform = 'CloudinaryLaravel/' . self::PACKAGE_VERSION;
+    }
+
+    /**
+    * Set Analytics
+    */
+    public function setAnalytics(): void
+    {
+        Analytics::sdkCode('W');
+        Analytics::sdkVersion(self::PACKAGE_VERSION);
+        Analytics::techVersion(app()->version());
     }
 
     /**
      * Create a Cloudinary Instance
      *
      */
-    public function bootCloudinary()
+    public function bootCloudinary(): void
     {
-        $this->cloudinary = new Cloudinary($this->cloudinaryConfig);
+        $this->cloudinary = new Cloudinary($this->url);
     }
 
     /**
      * Expose the Cloudinary Admin Functionality
      *
      */
-    public function admin()
+    public function admin(): AdminApi
     {
         return $this->cloudinary->adminApi();
     }
@@ -118,7 +136,7 @@ class CloudinaryEngine
      * Expose the Cloudinary Search Functionality
      *
      */
-    public function search()
+    public function search(): SearchApi
     {
         return $this->cloudinary->searchApi();
     }
@@ -136,13 +154,13 @@ class CloudinaryEngine
      * @param string $file The asset to upload.
      * @param array $options The optional parameters. See the upload API documentation.
      *
-     * @return ApiResponse
+     * @return CloudinaryEngine
      *
      * @throws ApiError
      *
      * @see https://cloudinary.com/documentation/image_upload_api_reference#upload_method
      */
-    public function upload($file, $options = [])
+    public function upload(string $file, array $options = []): static
     {
         $this->response = $this->uploadApi()->upload($file, $options);
 
@@ -153,26 +171,9 @@ class CloudinaryEngine
      * Expose the Cloudinary Upload Functionality
      *
      */
-    public function uploadApi()
+    public function uploadApi(): UploadApi
     {
         return $this->cloudinary->uploadApi();
-    }
-
-    /**
-     * Uploads an asset to a Cloudinary account.
-     *
-     * The asset can be:
-     * * a local file path
-     * * the actual data (byte array buffer)
-     * * the Data URI (Base64 encoded), max ~60 MB (62,910,000 chars)
-     * * the remote FTP, HTTP or HTTPS URL address of an existing file
-     * * a private storage bucket (S3 or Google Storage) URL of a whitelisted bucket
-     *
-     *  This is asynchronous
-     */
-    public function uploadAsync($file, $options = [])
-    {
-        return $this->uploadApi()->uploadAsync($file, $options);
     }
 
     /**
@@ -184,13 +185,13 @@ class CloudinaryEngine
      * @param string $uploadPreset The name of an upload preset.
      * @param array $options The optional parameters. See the upload API documentation.
      *
-     * @return ApiResponse
+     * @return CloudinaryEngine
      *
      * @throws ApiError
      *
      * @see https://cloudinary.com/documentation/image_upload_api_reference#unsigned_upload_syntax
      */
-    public function unsignedUpload($file, $uploadPreset, $options = [])
+    public function unsignedUpload(string $file, string $uploadPreset, array $options = []): static
     {
         $this->response = $this->uploadApi()->unsignedUpload($file, $uploadPreset, $options);
 
@@ -203,8 +204,9 @@ class CloudinaryEngine
      * The upload is not signed so an upload preset is required.
      *
      * This is asynchronous
+     * @throws ApiError
      */
-    public function unsignedUploadAsync($file, $uploadPreset, $options = [])
+    public function unsignedUploadAsync($file, $uploadPreset, $options = []): PromiseInterface
     {
         return $this->uploadApi()->unsignedUploadAsync($file, $uploadPreset, $options);
     }
@@ -215,7 +217,7 @@ class CloudinaryEngine
      * @return $this
      * @throws ApiError
      */
-    public function uploadFile($file, $options = [])
+    public function uploadFile($file, array $options = []): static
     {
         $uploadOptions = array_merge($options, ['resource_type' => 'auto']);
 
@@ -230,7 +232,7 @@ class CloudinaryEngine
      * @return $this
      * @throws ApiError
      */
-    public function uploadVideo($file, $options = [])
+    public function uploadVideo($file, array $options = []): static
     {
         $videoUploadOptions = array_merge($options, ['resource_type' => 'video']);
 
@@ -240,9 +242,9 @@ class CloudinaryEngine
     }
 
     /**
-     * @return Array
+     * @return array
      */
-    public function getResponse()
+    public function getResponse(): array
     {
         return $this->response;
     }
@@ -250,7 +252,7 @@ class CloudinaryEngine
     /**
      * @return mixed
      */
-    public function getAssetId()
+    public function getAssetId(): mixed
     {
         return $this->response[self::ASSET_ID];
     }
@@ -259,7 +261,7 @@ class CloudinaryEngine
      * Get the name of the file after it has been uploaded to Cloudinary
      * @return string
      */
-    public function getFileName()
+    public function getFileName(): string
     {
         return $this->response[self::PUBLIC_ID];
     }
@@ -268,7 +270,7 @@ class CloudinaryEngine
      * Get the public id of the file (also known as the name of the file) after it has been uploaded to Cloudinary
      * @return string
      */
-    public function getPublicId()
+    public function getPublicId(): string
     {
         return $this->response[self::PUBLIC_ID];
     }
@@ -277,7 +279,7 @@ class CloudinaryEngine
      * Get the name of the file before it was uploaded to Cloudinary
      * @return string
      */
-    public function getOriginalFileName()
+    public function getOriginalFileName(): string
     {
         return $this->response[self::ORIGINAL_FILENAME];
     }
@@ -285,7 +287,7 @@ class CloudinaryEngine
     /**
      * @return mixed
      */
-    public function getVersion()
+    public function getVersion(): mixed
     {
         return $this->response[self::VERSION];
     }
@@ -293,7 +295,7 @@ class CloudinaryEngine
     /**
      * @return mixed
      */
-    public function getVersionId()
+    public function getVersionId(): mixed
     {
         return $this->response[self::VERSION_ID];
     }
@@ -301,7 +303,7 @@ class CloudinaryEngine
     /**
      * @return mixed
      */
-    public function getSignature()
+    public function getSignature(): mixed
     {
         return $this->response[self::SIGNATURE];
     }
@@ -309,7 +311,7 @@ class CloudinaryEngine
     /**
      * @return mixed
      */
-    public function getWidth()
+    public function getWidth(): mixed
     {
         return $this->response[self::WIDTH];
     }
@@ -317,7 +319,7 @@ class CloudinaryEngine
     /**
      * @return mixed
      */
-    public function getHeight()
+    public function getHeight(): mixed
     {
         return $this->response[self::HEIGHT];
     }
@@ -325,7 +327,7 @@ class CloudinaryEngine
     /**
      * @return mixed
      */
-    public function getExtension()
+    public function getExtension(): mixed
     {
         return $this->response[self::FORMAT];
     }
@@ -333,7 +335,7 @@ class CloudinaryEngine
     /**
      * @return mixed
      */
-    public function getFileType()
+    public function getFileType(): mixed
     {
         return $this->response[self::RESOURCE_TYPE];
     }
@@ -341,7 +343,7 @@ class CloudinaryEngine
     /**
      * @return mixed
      */
-    public function getTimeUploaded()
+    public function getTimeUploaded(): mixed
     {
         return $this->response[self::CREATED_AT];
     }
@@ -349,7 +351,7 @@ class CloudinaryEngine
     /**
      * @return mixed
      */
-    public function getTags()
+    public function getTags(): mixed
     {
         return $this->response[self::TAGS];
     }
@@ -357,7 +359,7 @@ class CloudinaryEngine
     /**
      * @return mixed
      */
-    public function getPages()
+    public function getPages(): mixed
     {
         return $this->response[self::PAGES];
     }
@@ -365,7 +367,7 @@ class CloudinaryEngine
     /**
      * @return string
      */
-    public function getReadableSize()
+    public function getReadableSize(): string
     {
         return $this->getHumanReadableSize($this->getSize());
     }
@@ -373,10 +375,10 @@ class CloudinaryEngine
     /**
      * Formats filesize in the way every human understands
      *
-     * @param file $file
+     * @param $sizeInBytes
      * @return string Formatted Filesize, e.g. "113.24 MB".
      */
-    private function getHumanReadableSize($sizeInBytes)
+    public function getHumanReadableSize($sizeInBytes): string
     {
         if ($sizeInBytes >= 1073741824) {
             return number_format($sizeInBytes / 1073741824, 2) . ' GB';
@@ -396,7 +398,7 @@ class CloudinaryEngine
     /**
      * @return mixed
      */
-    public function getSize()
+    public function getSize(): mixed
     {
         return $this->response[self::BYTES];
     }
@@ -404,7 +406,7 @@ class CloudinaryEngine
     /**
      * @return mixed
      */
-    public function getPlaceHolder()
+    public function getPlaceHolder(): mixed
     {
         return $this->response[self::PLACEHOLDER];
     }
@@ -412,7 +414,7 @@ class CloudinaryEngine
     /**
      * @return mixed
      */
-    public function getPath()
+    public function getPath(): mixed
     {
         return $this->response[self::URL];
     }
@@ -420,7 +422,7 @@ class CloudinaryEngine
     /**
      * @return mixed
      */
-    public function getSecurePath()
+    public function getSecurePath(): mixed
     {
         return $this->response[self::SECURE_URL];
     }
@@ -428,9 +430,17 @@ class CloudinaryEngine
     /**
      * @return mixed
      */
-    public function getPhash()
+    public function getPhash(): mixed
     {
         return $this->response[self::PHASH];
+    }
+
+    /**
+     * @return mixed
+     */
+    public function getEtag(): mixed
+    {
+        return $this->response[self::ETAG];
     }
 
     /**
@@ -440,7 +450,7 @@ class CloudinaryEngine
      *
      * @return Image
      */
-    public function getImage($publicId)
+    public function getImage(string $publicId): Image
     {
         return $this->cloudinary->image($publicId);
     }
@@ -452,7 +462,7 @@ class CloudinaryEngine
      *
      * @return Video
      */
-    public function getVideo($publicId)
+    public function getVideo(mixed $publicId): Video
     {
         return $this->cloudinary->video($publicId);
     }
@@ -464,7 +474,7 @@ class CloudinaryEngine
      *
      * @return File
      */
-    public function getFile($publicId)
+    public function getFile(mixed $publicId): File
     {
         return $this->cloudinary->raw($publicId);
     }
@@ -473,7 +483,7 @@ class CloudinaryEngine
      * @param $publicId
      * @return ImageTag
      */
-    public function getImageTag($publicId)
+    public function getImageTag($publicId): ImageTag
     {
         return $this->cloudinary->imageTag($publicId);
     }
@@ -482,7 +492,7 @@ class CloudinaryEngine
      * @param $publicId
      * @return VideoTag
      */
-    public function getVideoTag($publicId)
+    public function getVideoTag($publicId): VideoTag
     {
         return $this->cloudinary->videoTag($publicId);
     }
@@ -504,7 +514,7 @@ class CloudinaryEngine
      *
      * @see https://cloudinary.com/documentation/image_upload_api_reference#tags_method
      */
-    public function addTag($tag, $publicIds = [], $options = [])
+    public function addTag(string $tag, array $publicIds = [], array $options = []): ApiResponse
     {
         return $this->uploadApi()->addTag($tag, $publicIds, $options);
     }
@@ -514,7 +524,7 @@ class CloudinaryEngine
      *
      * This is an asynchronous function.
      */
-    public function addTagAsync($tag, $publicIds = [], $options = [])
+    public function addTagAsync($tag, $publicIds = [], $options = []): PromiseInterface
     {
         return $this->uploadApi()->addTagAsync($tag, $publicIds, $options);
     }
@@ -530,7 +540,7 @@ class CloudinaryEngine
      *
      * @see https://cloudinary.com/documentation/image_upload_api_reference#tags_method
      */
-    public function removeTag($tag, $publicIds = [], $options = [])
+    public function removeTag(string $tag, array|string $publicIds = [], array $options = []): ApiResponse
     {
         return $this->uploadApi()->removeTag($tag, $publicIds, $options);
     }
@@ -541,7 +551,7 @@ class CloudinaryEngine
      * This is an asynchronous function.
      *
      */
-    public function removeTagAsync($tag, $publicIds = [], $options = [])
+    public function removeTagAsync($tag, $publicIds = [], $options = []): PromiseInterface
     {
         return $this->uploadApi()->removeTagAsync($tag, $publicIds, $options);
     }
@@ -556,7 +566,7 @@ class CloudinaryEngine
      *
      * @see https://cloudinary.com/documentation/image_upload_api_reference#tags_method
      */
-    public function removeAllTags($publicIds = [], $options = [])
+    public function removeAllTags(array $publicIds = [], array $options = []): ApiResponse
     {
         return $this->uploadApi()->removeAllTags($publicIds, $options);
     }
@@ -567,7 +577,7 @@ class CloudinaryEngine
      * This is an asynchronous function.
      *
      */
-    public function removeAllTagsAsync($publicIds = [], $options = [])
+    public function removeAllTagsAsync($publicIds = [], $options = []): PromiseInterface
     {
         return $this->uploadApi()->removeAllTagsAsync($publicIds, $options);
     }
@@ -583,7 +593,7 @@ class CloudinaryEngine
      *
      * @see https://cloudinary.com/documentation/image_upload_api_reference#tags_method
      */
-    public function replaceTag($tag, $publicIds = [], $options = [])
+    public function replaceTag(string $tag, array|string $publicIds = [], array $options = []): ApiResponse
     {
         return $this->uploadApi()->replaceTag($tag, $publicIds, $options);
     }
@@ -594,7 +604,7 @@ class CloudinaryEngine
      * This is an asynchronous function.
      *
      */
-    public function replaceTagAsync($tag, $publicIds = [], $options = [])
+    public function replaceTagAsync($tag, $publicIds = [], $options = []): PromiseInterface
     {
         return $this->uploadApi()->replaceTagAsync($tag, $publicIds, $options);
     }
@@ -619,7 +629,7 @@ class CloudinaryEngine
      *
      * @see https://cloudinary.com/documentation/image_upload_api_reference#sprite_method
      */
-    public function generateSprite($tag, $options = [])
+    public function generateSprite(string $tag, array $options = []): ApiResponse
     {
         return $this->uploadApi()->generateSprite($tag, $options);
     }
@@ -629,7 +639,7 @@ class CloudinaryEngine
      *
      * This is an asynchronous function.
      */
-    public function generateSpriteAsync($tag, $options = [])
+    public function generateSpriteAsync($tag, $options = []): PromiseInterface
     {
         return $this->uploadApi()->generateSpriteAsync($tag, $options);
     }
@@ -644,7 +654,7 @@ class CloudinaryEngine
      *
      * @see https://cloudinary.com/documentation/paged_and_layered_media#creating_pdf_files_from_images
      */
-    public function generatePDF($tag, $options = [])
+    public function generatePDF($tag, $options = []): ApiResponse
     {
         $pdfOptions = array_merge($options, ['async' => false, 'format' => 'pdf']);
 
@@ -656,7 +666,7 @@ class CloudinaryEngine
      * @param array $options
      * @return ApiResponse
      */
-    public function generatePDFAsync($tag, $options = [])
+    public function generatePDFAsync($tag, array $options = []): ApiResponse
     {
         $pdfOptions = array_merge($options, ['async' => true, 'format' => 'pdf']);
 
@@ -668,7 +678,7 @@ class CloudinaryEngine
      * @param array $options
      * @return ApiResponse
      */
-    public function generateAnimatedGIF($tag, $options = [])
+    public function generateAnimatedGIF($tag, array $options = []): ApiResponse
     {
         $gifOptions = array_merge($options, ['async' => false, 'format' => 'gif']);
 
@@ -680,7 +690,7 @@ class CloudinaryEngine
      * @param array $options
      * @return ApiResponse
      */
-    public function generateAnimatedPNG($tag, $options = [])
+    public function generateAnimatedPNG($tag, array $options = []): ApiResponse
     {
         $pngOptions = array_merge($options, ['async' => false, 'format' => 'png']);
 
@@ -692,7 +702,7 @@ class CloudinaryEngine
      * @param array $options
      * @return ApiResponse
      */
-    public function generateAnimatedPNGAsync($tag, $options = [])
+    public function generateAnimatedPNGAsync($tag, array $options = []): ApiResponse
     {
         $pngOptions = array_merge($options, ['async' => true, 'format' => 'png']);
 
@@ -704,7 +714,7 @@ class CloudinaryEngine
      * @param array $options
      * @return ApiResponse
      */
-    public function generateAnimatedWEBP($tag, $options = [])
+    public function generateAnimatedWEBP($tag, array $options = []): ApiResponse
     {
         $webpOptions = array_merge($options, ['async' => false, 'format' => 'webp']);
 
@@ -716,7 +726,7 @@ class CloudinaryEngine
      * @param array $options
      * @return ApiResponse
      */
-    public function generateAnimatedWEBPAsync($tag, $options = [])
+    public function generateAnimatedWEBPAsync($tag, array $options = []): ApiResponse
     {
         $webpOptions = array_merge($options, ['async' => true, 'format' => 'webp']);
 
@@ -728,7 +738,7 @@ class CloudinaryEngine
      * @param array $options
      * @return ApiResponse
      */
-    public function generateAnimatedMP4($tag, $options = [])
+    public function generateAnimatedMP4($tag, array $options = []): ApiResponse
     {
         $mp4Options = array_merge($options, ['async' => false, 'format' => 'mp4']);
 
@@ -740,7 +750,7 @@ class CloudinaryEngine
      * @param array $options
      * @return ApiResponse
      */
-    public function generateAnimatedMP4Async($tag, $options = [])
+    public function generateAnimatedMP4Async($tag, array $options = []): ApiResponse
     {
         $mp4Options = array_merge($options, ['async' => true, 'format' => 'mp4']);
 
@@ -752,7 +762,7 @@ class CloudinaryEngine
      * @param array $options
      * @return ApiResponse
      */
-    public function generateAnimatedWEBM($tag, $options = [])
+    public function generateAnimatedWEBM($tag, array $options = []): ApiResponse
     {
         $webmOptions = array_merge($options, ['async' => false, 'format' => 'webm']);
 
@@ -764,7 +774,7 @@ class CloudinaryEngine
      * @param array $options
      * @return ApiResponse
      */
-    public function generateAnimatedWEBMAsync($tag, $options = [])
+    public function generateAnimatedWEBMAsync($tag, array $options = []): ApiResponse
     {
         $webmOptions = array_merge($options, ['async' => true, 'format' => 'webm']);
 
@@ -776,7 +786,7 @@ class CloudinaryEngine
      * @param array $options
      * @return ApiResponse
      */
-    public function multi($tag, $options = [])
+    public function multi($tag, array $options = []): ApiResponse
     {
         return $this->uploadApi()->multi($tag, $options);
     }
@@ -786,7 +796,7 @@ class CloudinaryEngine
      * @param array $options
      * @return PromiseInterface
      */
-    public function multiAsync($tag, $options = [])
+    public function multiAsync($tag, array $options = []): PromiseInterface
     {
         return $this->uploadApi()->multiAsync($tag, $options);
     }
@@ -796,7 +806,7 @@ class CloudinaryEngine
      * @param array $options
      * @return ApiResponse
      */
-    public function explode($publicId, $options = [])
+    public function explode($publicId, array $options = []): ApiResponse
     {
         return $this->uploadApi()->explode($publicId, $options);
     }
@@ -806,7 +816,7 @@ class CloudinaryEngine
      * @param array $options
      * @return PromiseInterface
      */
-    public function explodeAsync($publicId, $options = [])
+    public function explodeAsync($publicId, array $options = []): PromiseInterface
     {
         return $this->uploadApi()->explodeAsync($publicId, $options);
     }
@@ -817,11 +827,11 @@ class CloudinaryEngine
      * @param string $text The text string to generate an image for.
      * @param array $options The optional parameters.  See the upload API documentation.
      *
-     * @return ApiResponse
+     * @return CloudinaryEngine
      *
      * @see https://cloudinary.com/documentation/image_upload_api_reference#text_method
      */
-    public function generateImageFromText($text, $options = [])
+    public function generateImageFromText(string $text, array $options = []): static
     {
         $this->response = $this->uploadApi()->text($text, $options);
 
@@ -833,7 +843,7 @@ class CloudinaryEngine
      * @param array $options
      * @return PromiseInterface
      */
-    public function generateImageFromTextAsync($text, $options = [])
+    public function generateImageFromTextAsync($text, array $options = []): PromiseInterface
     {
         return $this->uploadApi()->textAsync($text, $options);
     }
@@ -849,7 +859,7 @@ class CloudinaryEngine
      * @param null $targetFormat
      * @return ApiResponse
      */
-    public function createArchive($options = [], $targetFormat = null)
+    public function createArchive(array $options = [], $targetFormat = null): ApiResponse
     {
         return $this->uploadApi()->createArchive($options, $targetFormat);
     }
@@ -859,7 +869,7 @@ class CloudinaryEngine
      * @param null $targetFormat
      * @return PromiseInterface
      */
-    public function createArchiveAsync($options = [], $targetFormat = null)
+    public function createArchiveAsync(array $options = [], $targetFormat = null): PromiseInterface
     {
         return $this->uploadApi()->createArchiveAsync($options, $targetFormat);
     }
@@ -868,7 +878,7 @@ class CloudinaryEngine
      * @param array $options
      * @return ApiResponse
      */
-    public function createZip($options = [])
+    public function createZip(array $options = []): ApiResponse
     {
         return $this->uploadApi()->createZip($options);
     }
@@ -877,7 +887,7 @@ class CloudinaryEngine
      * @param array $options
      * @return PromiseInterface
      */
-    public function createZipAsync($options = [])
+    public function createZipAsync(array $options = []): PromiseInterface
     {
         return $this->uploadApi()->createZipAsync($options);
     }
@@ -886,7 +896,7 @@ class CloudinaryEngine
      * @param array $options
      * @return string
      */
-    public function downloadZipUrl($options = [])
+    public function downloadZipUrl(array $options = []): string
     {
         return $this->uploadApi()->downloadZipUrl($options);
     }
@@ -895,7 +905,7 @@ class CloudinaryEngine
      * @param array $options
      * @return string
      */
-    public function downloadArchiveUrl($options = [])
+    public function downloadArchiveUrl(array $options = []): string
     {
         return $this->uploadApi()->downloadArchiveUrl($options);
     }
@@ -912,7 +922,7 @@ class CloudinaryEngine
      * @param array $options
      * @return ApiResponse
      */
-    public function addContext($context, $publicIds = [], $options = [])
+    public function addContext($context, array $publicIds = [], array $options = []): ApiResponse
     {
         return $this->uploadApi()->addContext($context, $publicIds, $options);
     }
@@ -923,7 +933,7 @@ class CloudinaryEngine
      * @param array $options
      * @return PromiseInterface
      */
-    public function addContextAsync($context, $publicIds = [], $options = [])
+    public function addContextAsync($context, array $publicIds = [], array $options = []): PromiseInterface
     {
         return $this->uploadApi()->addContextAsync($context, $publicIds, $options);
     }
@@ -933,7 +943,7 @@ class CloudinaryEngine
      * @param array $options
      * @return ApiResponse
      */
-    public function removeAllContext($publicIds = [], $options = [])
+    public function removeAllContext(array $publicIds = [], array $options = []): ApiResponse
     {
         return $this->uploadApi()->removeAllContext($publicIds, $options);
     }
@@ -943,7 +953,7 @@ class CloudinaryEngine
      * @param array $options
      * @return PromiseInterface
      */
-    public function removeAllContextAsync($publicIds = [], $options = [])
+    public function removeAllContextAsync(array $publicIds = [], array $options = []): PromiseInterface
     {
         return $this->uploadApi()->removeAllContextAsync($publicIds, $options);
     }
@@ -957,9 +967,9 @@ class CloudinaryEngine
     /**
      * @param $publicId
      * @param array $options
-     * @return ApiResponse
+     * @return array|ApiResponse
      */
-    public function destroy($publicId, $options = [])
+    public function destroy($publicId, array $options = []): array|ApiResponse
     {
         return $this->uploadApi()->destroy($publicId, $options);
     }
@@ -969,7 +979,7 @@ class CloudinaryEngine
      * @param array $options
      * @return PromiseInterface
      */
-    public function destroyAsync($publicId, $options = [])
+    public function destroyAsync($publicId, array $options = []): PromiseInterface
     {
         return $this->uploadApi()->destroyAsync($publicId, $options);
     }
@@ -980,7 +990,7 @@ class CloudinaryEngine
      * @param array $options
      * @return mixed
      */
-    public function rename($from, $to, $options = [])
+    public function rename($from, $to, array $options = []): mixed
     {
         return $this->uploadApi()->rename($from, $to, $options);
     }
@@ -991,7 +1001,7 @@ class CloudinaryEngine
      * @param array $options
      * @return PromiseInterface
      */
-    public function renameAsync($from, $to, $options = [])
+    public function renameAsync($from, $to, array $options = []): PromiseInterface
     {
         return $this->uploadApi()->renameAsync($from, $to, $options);
     }
@@ -1001,7 +1011,7 @@ class CloudinaryEngine
      * @param array $options
      * @return mixed
      */
-    public function explicit($publicId, $options = [])
+    public function explicit($publicId, array $options = []): mixed
     {
         return $this->uploadApi()->explicit($publicId, $options);
     }
@@ -1011,7 +1021,7 @@ class CloudinaryEngine
      * @param array $options
      * @return PromiseInterface
      */
-    public function explicitAsync($publicId, $options = [])
+    public function explicitAsync($publicId, array $options = []): PromiseInterface
     {
         return $this->uploadApi()->explicitAsync($publicId, $options);
     }
@@ -1019,13 +1029,13 @@ class CloudinaryEngine
     /**
      * Get Resource data
      * @param string $path
-     * @return array
+     * @return ApiResponse|string;
      */
-    public function getResource($path)
+    public function getResource(string $path): string|ApiResponse
     {
         try {
             return $this->admin()->asset($path);
-        } catch (\Exception $e) {
+        } catch (Exception) {
             return '';
         }
     }
@@ -1037,7 +1047,7 @@ class CloudinaryEngine
      *
      * @return string|false
      */
-    public function getUrl($publicId)
+    public function getUrl(string $publicId): bool|string
     {
 
         $resource = $this->getResource($publicId);