+++ src/CloudinaryServiceProvider.php
@@ -8,6 +8,7 @@ use Illuminate\Support\Facades\Route;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\ServiceProvider;
 use League\Flysystem\Filesystem;
+use Illuminate\Filesystem\FilesystemAdapter;
 use CloudinaryLabs\CloudinaryLaravel\Commands\BackupFilesCommand;
 use CloudinaryLabs\CloudinaryLaravel\Commands\DeleteFilesCommand;
 use CloudinaryLabs\CloudinaryLaravel\Commands\FetchFilesCommand;
@@ -27,7 +28,7 @@ class CloudinaryServiceProvider extends ServiceProvider
      *
      * @return void
      */
-    public function boot()
+    public function boot(): void
     {
         $this->bootMacros();
         $this->bootResources();
@@ -44,7 +45,7 @@ class CloudinaryServiceProvider extends ServiceProvider
      *
      * @return void
      */
-    protected function bootMacros()
+    protected function bootMacros(): void
     {
         UploadedFile::macro(
             'storeOnCloudinary',
@@ -69,7 +70,7 @@ class CloudinaryServiceProvider extends ServiceProvider
      *
      * @return void
      */
-    protected function bootResources()
+    protected function bootResources(): void
     {
         $this->loadViewsFrom(__DIR__ . '/../resources/views', 'cloudinary');
     }
@@ -79,7 +80,7 @@ class CloudinaryServiceProvider extends ServiceProvider
      *
      * @return void
      */
-    protected function bootDirectives()
+    protected function bootDirectives(): void
     {
         Blade::directive(
             'cloudinaryJS',
@@ -94,7 +95,7 @@ class CloudinaryServiceProvider extends ServiceProvider
      *
      * @return void
      */
-    protected function bootComponents()
+    protected function bootComponents(): void
     {
         Blade::component('cloudinary::components.button', $this->getComponentName('cld-upload-button'));
         Blade::component('cloudinary::components.image', $this->getComponentName('cld-image'));
@@ -107,11 +108,11 @@ class CloudinaryServiceProvider extends ServiceProvider
        if( (int)$version[0] <= 6 ) {
           $componentName = str_replace("-", "_", $componentName);
        }
-        
+
        return $componentName;
     }
 
-    protected function bootCommands()
+    protected function bootCommands(): void
     {
         /**
          * Register Laravel Cloudinary Artisan commands
@@ -135,7 +136,7 @@ class CloudinaryServiceProvider extends ServiceProvider
      *
      * @return void
      */
-    protected function bootPublishing()
+    protected function bootPublishing(): void
     {
         if ($this->app->runningInConsole()) {
             $config = dirname(__DIR__) . '/config/cloudinary.php';
@@ -156,14 +157,21 @@ class CloudinaryServiceProvider extends ServiceProvider
         }
     }
 
-    protected function bootCloudinaryDriver()
+    protected function bootCloudinaryDriver(): void
     {
         $this->app['config']['filesystems.disks.cloudinary'] = ['driver' => 'cloudinary'];
 
         Storage::extend(
             'cloudinary',
             function ($app, $config) {
-                return new Filesystem(new CloudinaryAdapter(config('cloudinary.cloud_url')));
+
+                $cloudinaryAdapter = new CloudinaryAdapter(config('cloudinary.cloud_url'));
+
+                return new FilesystemAdapter(
+                    new Filesystem($cloudinaryAdapter, $config),
+                    $cloudinaryAdapter,
+                    $config
+                );
             }
         );
     }
@@ -173,7 +181,7 @@ class CloudinaryServiceProvider extends ServiceProvider
      *
      * @return void
      */
-    protected function bootRoutes()
+    protected function bootRoutes(): void
     {
         if (config('cloudinary.upload_route')) {
             Route::post(config('cloudinary.upload_route'), config('cloudinary.upload_action'));
@@ -185,12 +193,12 @@ class CloudinaryServiceProvider extends ServiceProvider
      *
      * @return void
      */
-    public function register()
+    public function register(): void
     {
         // Register the service the package provides.
         $this->app->singleton(
             CloudinaryEngine::class,
-            function ($app) {
+            function () {
                 return new CloudinaryEngine();
             }
         );