+++ src/Facades/Cloudinary.php
@@ -16,7 +16,7 @@ class Cloudinary extends Facade
      *
      * @return string
      */
-    protected static function getFacadeAccessor()
+    protected static function getFacadeAccessor(): string
     {
         return CloudinaryEngine::class;
     }