+++ changelog.md
@@ -35,3 +35,27 @@
   * Fix Flystem Adapter for Cloudinary Adapter
   * Add support for getting asset url using only public id
   * Replace deprecated method names with v2
+
+2.0.0 / 2022-06-02
+======================
+
+  * Rewrite Cloudinary Adapter to work with Flysystem v3
+  * Support for Laravel 9 without issues
+  * Remove deprecated methods
+
+2.0.1 / 2022-06-03
+======================
+
+  * Fix listContents() and make it compatible with Flysystem v3. @brandon14
+
+
+2.0.2 / 2023-02-07
+======================
+
+  * Update Analytics
+
+
+2.0.3 / 2023-03-01
+======================
+
+  * Bump to support Laravel 10